final_project/requirements.txt.txt
streamlit
pandas
matplotlib.pyplot
pydeck
PIL